Feat: Test Harness (#1)
* fix(fmt): apply

* fix(foundry): tests

* fix error

+++ src/IntentionLockLibrary.sol
@@ -10,18 +10,24 @@
 pragma solidity ^0.8.21;
 
 contract IntentionLock {
-    enum LockState {Unlocked, X, S, IX, IS}
-    
+    enum LockState {
+        Unlocked,
+        X,
+        S,
+        IX,
+        IS
+    }
+
     struct Node {
         LockState state;
         address owner;
-        uint parentIndex;
+        uint256 parentIndex;
     }
 
     Node[] public tree;
 
     // Conditions
-    function _isOwner(uint nodeIndex, address user) private view returns (bool) {
+    function _isOwner(uint256 nodeIndex, address user) private view returns (bool) {
         return tree[nodeIndex].owner == user;
     }
 
@@ -38,41 +44,42 @@ contract IntentionLock {
     }
 
     function _canLockIS(LockState currentState) private pure returns (bool) {
-        return currentState == LockState.Unlocked || currentState == LockState.S || currentState == LockState.IS || currentState == LockState.IX;
+        return currentState == LockState.Unlocked || currentState == LockState.S || currentState == LockState.IS
+            || currentState == LockState.IX;
     }
-    
+
     function _isRootNodeExist() private view returns (bool) {
         return tree.length > 0;
     }
 
-    function _isValidParent(uint parentIndex) private view returns (bool) {
+    function _isValidParent(uint256 parentIndex) private view returns (bool) {
         return parentIndex < tree.length;
     }
 
     // Modifiers
     // ... [function modifiers] ...
 
-    modifier onlyOwner(uint nodeIndex) {
+    modifier onlyOwner(uint256 nodeIndex) {
         require(tree[nodeIndex].owner == msg.sender, "Not the owner");
         _;
     }
 
-    modifier canLockX(uint nodeIndex) {
+    modifier canLockX(uint256 nodeIndex) {
         require(_canLockX(tree[nodeIndex].state), "Cannot lock X");
         _;
     }
 
-    modifier canLockS(uint nodeIndex) {
+    modifier canLockS(uint256 nodeIndex) {
         require(_canLockS(tree[nodeIndex].state), "Cannot lock S");
         _;
     }
 
-    modifier canLockIX(uint nodeIndex) {
+    modifier canLockIX(uint256 nodeIndex) {
         require(_canLockIX(tree[nodeIndex].state), "Cannot lock IX");
         _;
     }
 
-    modifier canLockIS(uint nodeIndex) {
+    modifier canLockIS(uint256 nodeIndex) {
         require(_canLockIS(tree[nodeIndex].state), "Cannot lock IS");
         _;
     }
@@ -82,7 +89,7 @@ contract IntentionLock {
         _;
     }
 
-    modifier isValidParent(uint parentIndex) {
+    modifier isValidParent(uint256 parentIndex) {
         require(_isValidParent(parentIndex), "Invalid parent index");
         _;
     }
@@ -91,11 +98,7 @@ contract IntentionLock {
 
     // _addRootNode
     function _addRootNode() private {
-        Node memory newNode = Node({
-            state: LockState.Unlocked,
-            owner: address(0),
-            parentIndex: 0 
-        });
+        Node memory newNode = Node({state: LockState.Unlocked, owner: address(0), parentIndex: 0});
         tree.push(newNode);
         emit NodeAdded(0, 0, msg.sender);
 
@@ -103,21 +106,17 @@ contract IntentionLock {
         assert(tree[0].parentIndex == 0 && tree.length == 1);
     }
 
-    function _addChildNode(uint parentIndex) private {
-        Node memory newNode = Node({
-            state: LockState.Unlocked,
-            owner: address(0),
-            parentIndex: parentIndex
-        });
+    function _addChildNode(uint256 parentIndex) private {
+        Node memory newNode = Node({state: LockState.Unlocked, owner: address(0), parentIndex: parentIndex});
         tree.push(newNode);
-        uint newNodeIndex = tree.length - 1;
+        uint256 newNodeIndex = tree.length - 1;
         emit NodeAdded(newNodeIndex, parentIndex, msg.sender);
 
         // Assertion for invariant: Child node's parent index matches the provided parent index
         assert(tree[newNodeIndex].parentIndex == parentIndex);
     }
 
-    function _lockNode(uint nodeIndex, LockState lockState) private {
+    function _lockNode(uint256 nodeIndex, LockState lockState) private {
         tree[nodeIndex].state = lockState;
         tree[nodeIndex].owner = msg.sender;
         emit NodeLocked(nodeIndex, lockState, msg.sender);
@@ -126,7 +125,7 @@ contract IntentionLock {
         assert(tree[nodeIndex].state == lockState && tree[nodeIndex].owner == msg.sender);
     }
 
-    function _unlockNode(uint nodeIndex) private {
+    function _unlockNode(uint256 nodeIndex) private {
         tree[nodeIndex].state = LockState.Unlocked;
         tree[nodeIndex].owner = address(0);
         emit NodeUnlocked(nodeIndex, msg.sender);
@@ -136,9 +135,9 @@ contract IntentionLock {
     }
 
     // Event Emit
-    event NodeAdded(uint nodeIndex, uint parentIndex, address createdBy);
-    event NodeLocked(uint nodeIndex, LockState lockedState, address lockedBy);
-    event NodeUnlocked(uint nodeIndex, address unlockedBy);
+    event NodeAdded(uint256 nodeIndex, uint256 parentIndex, address createdBy);
+    event NodeLocked(uint256 nodeIndex, LockState lockedState, address lockedBy);
+    event NodeUnlocked(uint256 nodeIndex, address unlockedBy);
 
     // Combined Functions
     // ... [combined functions] ...
@@ -147,27 +146,27 @@ contract IntentionLock {
         _addRootNode();
     }
 
-    function addChildNode(uint parentIndex) public isValidParent(parentIndex) {
+    function addChildNode(uint256 parentIndex) public isValidParent(parentIndex) {
         _addChildNode(parentIndex);
     }
 
-    function lockX(uint nodeIndex) public canLockX(nodeIndex) {
+    function lockX(uint256 nodeIndex) public canLockX(nodeIndex) {
         _lockNode(nodeIndex, LockState.X);
     }
 
-    function lockS(uint nodeIndex) public canLockS(nodeIndex) {
+    function lockS(uint256 nodeIndex) public canLockS(nodeIndex) {
         _lockNode(nodeIndex, LockState.S);
     }
 
-    function lockIX(uint nodeIndex) public canLockIX(nodeIndex) {
+    function lockIX(uint256 nodeIndex) public canLockIX(nodeIndex) {
         _lockNode(nodeIndex, LockState.IX);
     }
 
-    function lockIS(uint nodeIndex) public canLockIS(nodeIndex) {
+    function lockIS(uint256 nodeIndex) public canLockIS(nodeIndex) {
         _lockNode(nodeIndex, LockState.IS);
     }
 
-    function unlock(uint nodeIndex) public onlyOwner(nodeIndex) {
+    function unlock(uint256 nodeIndex) public onlyOwner(nodeIndex) {
         _unlockNode(nodeIndex);
     }
 }
